Refactor the frontend into its own module
